requirements.txt
Django==4.2
djangorestframework==3.15.2
requests==2.32.3
python-dateutil==2.9.0
gunicorn==23.0.0
pytest==8.3.4
pytest-django==4.10.0

# Slack API
slack_sdk==3.35.0

# Azure AI Search
azure-search-documents==11.5.2

# Azure OpenAI
openai==1.76.2

# Document processing
python-docx==1.1.2
PyPDF2==3.0.1
beautifulsoup4==4.13.4

# Excel processing
pandas==2.2.2
openpyxl==2.0.0
